requirements.txt
panda==2.2.3
openpyxl==3.1.2
lxml==5.1.0
tk==0.1.0
ttkthemes==3.2.2
# pyinstaller==6.12.0